Locking seems to work now on both AFS and NFSv3 with reasonable speed

+++ python/Ganga/Core/GangaRepository/SessionLock.py
@@ -1,7 +1,12 @@
 # This file is a bit of a mess at the moment.
 # It encapsulates fcntl-based locking that works on NFS, AFS and locally.
-import os, time, pickle, errno, threading, fcntl, random
+import os, time, errno, threading, fcntl, random
 from sets import Set
+
+try:
+    import cPickle as pickle
+except ImportError:
+    import pickle
 
 try:
     from Ganga.Core.GangaThread import GangaThread
@@ -44,124 +49,135 @@ class SessionLock(GangaThread):
     across clients, even if a global lock file is used!
     TODO: On OSError it should shutdown the repository!"""
 
-    def __init__(self, dir, session_name, name):
+    def __init__(self, dir, name):
         GangaThread.__init__(self, name='LockUpdater.%s' % name)
         realpath = os.path.realpath(dir)
+        # Use the hostname (os.uname()[1])  and the current time in ms to construct the session filename.
+        # TODO: Perhaps put the username here?
+        session_name = os.uname()[1]+"."+str(int(time.time()*1000))+".session"
         self.sdir = os.path.join(realpath,"sessions")
         self.ldir = os.path.join(realpath,"locks")
-        mkdir(self.sdir)
-        mkdir(self.ldir)
         self.fn = os.path.join(self.sdir, session_name)
+        self.cntfn = os.path.join(realpath,"cnt")
+
         self.afs = (realpath[:4] == "/afs")
         self.locked = Set()
-        if self.afs:
-            self.afslockfn = os.path.join(self.ldir,"afslock")
-            file(self.afslockfn,"w").close() # create file
-            self.afslockfd = os.open(os.afslockfn,os.O_RDWR)            
+        self.count = 0
         self.create()
 
     def create(self):
+        mkdir(self.sdir)
+        mkdir(self.ldir)
+        lockfn = os.path.join(self.sdir,"global_lock")
+        file(lockfn,"w").close() # create file
+        self.lockfd = os.open(lockfn,os.O_RDWR)
+        if not os.path.exists(self.cntfn):
+            try:
+                fd = os.open(self.cntfn, os.O_EXCL | os.O_CREAT | os.O_WRONLY)
+                os.write(fd,"0")
+                os.close(fd)
+            except OSError, x:
+                if x.errno != errno.EEXIST:
+                    raise RepositoryError("OSError on count file create: %s" % x)
+        else:
+            f = file(self.cntfn)
+            try:
+                self.count = max(self.count,int(f.readline()))
+            except ValueError:
+                pass
+            f.close()
         try:
-            fd = os.open(self.fn, os.O_EXCL | os.O_CREAT)
+            fd = os.open(self.fn, os.O_EXCL | os.O_CREAT | os.O_WRONLY)
+            os.write(fd,pickle.dumps(Set()))
             os.close(fd)
         except OSError, x:
             raise RepositoryError("Error on file access - session file: %s" % x)
 
-    def afs_global_lock(self):
-        fcntl.flock(self.afslockfd,fcntl.LOCK_EX)
+    def global_lock(self):
+        fcntl.lockf(self.lockfd,fcntl.LOCK_EX)
 
-    def afs_global_unlock(self):
-        fcntl.flock(self.afslockfd,fcntl.LOCK_UN)
+    def global_unlock(self):
+        fcntl.lockf(self.lockfd,fcntl.LOCK_UN)
 
-    def acquire(self,name):
-        if self.afs:
-            return self.acquire_afs(name)
-        else:
-            return self.acquire_nonafs(name)
-
-    def acquire_afs(self,name):
-        if name in self.locked:
-            return True
-        self.afs_global_lock()
+    def acquire_next(self,n):
+        self.global_lock()
         try:
-            sessions = os.listdir(self.sdir)
-            for session in sessions:
-                if name in file(os.path.join(self.sdir,session),"r").readlines():
-                    return False
-            f = file(self.fn,"w")
-            f.writelines("\n".join(self.locked))
-            f.close()
-            return True
-        finally:
-            self.afs_global_unlock()
-
-    def acquire_nonafs(self,name):
-        lfn = os.path.join(self.ldir,name)
-        try:
-            os.symlink(self.fn,lfn)
-            return True
-        except OSError:
-            pass
-        try:
-            if os.path.islink(lfn) and os.path.realpath(lfn) == self.fn:
-                return True # uh, this is already our lock...
-        except OSError:
-            return False # ok, this has gone south, just try again if you really want...
-        if os.path.exists(lfn): # Someone else has a valid lock
-            return False
-
-        # Now there is a broken link - this is somewhat uncomfortable, since
-        # we now have to avoid race conditions between two sessions.
-        # The solution is to use a dot-lock and only remove+recreate the symlink
-        # if the link is still broken after the dot-lock succeeds
-        # If this dot-lock stays there, we have a big problem. Therefore make sure it gets unlinked!
-        try:
-            fd = os.open(lfn+".lock", os.O_EXCL | os.O_CREAT | os.O_NONBLOCK)
+            # Actualize count
+            f = file(self.cntfn)
             try:
-                if not os.path.exists(lfn): # Recheck that this link is still invalid
-                    os.unlink(lfn)
-                    os.symlink(self.fn,lfn)
-            finally:
-                os.close(fd)
-                os.unlink(lfn+".lock")
-        except OSError: # did not get the dot-lock :(
-            pass
-        # ... someone else has the lock 
-        return False
+                newcount = int(f.readline())
+            except ValueError:
+                logger.warning("Corrupt job counter! Trying to recover...")
+                newcount = self.count
+            f.close()
+            assert newcount >= self.count
+            ids = range(newcount,newcount+n)
+            self.locked.update(ids)
+            self.count = newcount+n
+            f = file(self.cntfn,"w")
+            f.write(str(self.count))
+            f.close()
+            fd = os.open(self.fn,os.O_WRONLY)
+            fcntl.lockf(fd,fcntl.LOCK_EX) # ONLY FOR NFS
+            os.write(fd,pickle.dumps(self.locked))
+            fcntl.lockf(fd,fcntl.LOCK_UN) # ONLY FOR NFS
+            os.close(fd)
+            return ids
+        finally:
+            self.global_unlock()
+
+    def acquire(self,ids):
+        ids = Set(ids)
+        self.global_lock()
+        try:
+            sessions = [sn for sn in os.listdir(self.sdir) if sn.endswith(".session")]
+            slocked = Set()
+            for session in sessions:
+                sf = os.path.join(self.sdir,session)
+                if sf == self.fn:
+                    continue
+                fd = -1
+                if not self.afs:
+                    fd = os.open(sf, os.O_RDONLY)
+                    fcntl.lockf(fd,fcntl.LOCK_SH) # ONLY NFS
+                try:
+                    slocked.update(pickle.load(file(sf)))
+                except Exception:
+                    logger.warning("Corrupt session file - ignoring it: '%s'"% sf)
+                if not self.afs and fd > 0:
+                    fcntl.lockf(fd,fcntl.LOCK_UN) # ONLY NFS
+                    os.close(fd)
+            ids.difference_update(slocked)
+            self.locked.update(ids)
+            fd = os.open(self.fn,os.O_WRONLY)
+            fcntl.lockf(fd,fcntl.LOCK_EX) # ONLY FOR NFS
+            os.write(fd,pickle.dumps(self.locked))
+            fcntl.lockf(fd,fcntl.LOCK_UN) # ONLY FOR NFS
+            os.close(fd)
+            return ids
+        finally:
+            self.global_unlock()
 
     def acquire_block(self,name):
         while not self.acquire(name):
             time.sleep(0.01)
 
-    def release(self,name):
-        if self.afs:
-            self.release_afs(name)
-        else:
-            self.release_nonafs(name)
-
-    def release_afs(self,name):
-        self.afs_global_lock()
+    def release(self,ids):
+        self.global_lock()
         try:
-            self.locked.remove(name)
-            f = file(self.fn,"w")
-            f.writelines("\n".join(self.locked))
-            f.close()
+            self.locked.difference_update(ids)
+            fd = os.open(self.fn,os.O_WRONLY)
+            fcntl.lockf(fd,fcntl.LOCK_EX) # ONLY FOR NFS
+            os.write(fd,pickle.dumps(self.locked))
+            fcntl.lockf(fd,fcntl.LOCK_UN) # ONLY FOR NFS
+            os.close(fd)
+            return ids
         finally:
-            self.afs_global_unlock()
+            self.global_unlock()
 
-    def release_nonafs(self,name):
-        try:
-            if os.readlink(os.path.join(self.ldir,name)) == self.fn:
-                os.unlink(os.path.join(self.ldir,name))
-                return True
-        except IOError:
-            pass
-        return False
-        
     def run(self):
         while not self.should_stop():
             ## TODO: Check for services active/inactive
-            time.sleep(1+random.random())
             try:
                 if not os.path.exists(self.fn):
                     self.create()
@@ -172,6 +188,8 @@ class SessionLock(GangaThread):
                 try:
                     now = os.stat(self.fn).st_ctime
                     for sf in os.listdir(self.sdir):
+                        if not sf.endswith(".session"):
+                            continue
                         if now - os.stat(os.path.join(self.sdir,sf)).st_ctime > session_expiration_timeout:
                             os.unlink(os.path.join(self.sdir,sf))
                 except OSError, x:
@@ -179,13 +197,13 @@ class SessionLock(GangaThread):
                     logger.warning("Unimportant OSError in loop: %s" % x)
             except Exception, x:
                 logger.warning("Internal exception in session lock thread: %s %s" % (x.__class__.__name__, x))
+            time.sleep(1+random.random())
 
         # Remove session file
         try:
             os.unlink(self.fn)
             # close AFS lock file
-            if self.afs:
-                os.close(self.afslockfd)
+            os.close(self.lockfd)
         except OSError, x:
             logger.warning("Session file was deleted externally or removal failed: %s" % (x))
         self.unregister()
@@ -198,96 +216,67 @@ class SessionLockManager(object):
         This means that NO other part/thread/etc. of ganga should do ANY operations
         on these files as the locks will be lost!
         Raise RepositoryError"""
-
-        self.locked_ids = Set()
-        self.foreign_ids = []
         self.root = root_dir
         self.name = name
-
         mkdir(self.root)
-
-        self.cntfn = os.path.join(self.root, "count")
-        if not os.path.exists(self.cntfn):
-            try:
-                fd = os.open(self.cntfn, os.O_EXCL | os.O_CREAT | os.O_WRONLY)
-                os.write(fd,"0")
-                os.close(fd)
-            except OSError, x:
-                if x.errno != errno.EEXIST:
-                    raise RepositoryError("OSError on count file create: %s" % x)
-
-        # Use the hostname (os.uname()[1])  and the current time in ms to construct the session filename.
-        # TODO: Perhaps put the username here?
-        session_name = os.uname()[1]+"."+str(int(time.time()*1000))+".session"
-        self.sessionlock = SessionLock(self.root, session_name, self.name)
+        self.sessionlock = SessionLock(self.root, self.name)
         self.sessionlock.start()
         
     def make_new_ids(self,n):
         """ make_new_ids(n) --> list of ids
         Generate n consecutive new job ids.
         Raise RepositoryError"""
-        ids = []
-        self.sessionlock.acquire_block("counter")
-        try:
-            try:
-                cnt = int(file(self.cntfn).readlines()[0])
-                nl = 0
-                while nl < n:
-                    if self.sessionlock.acquire(str(cnt)):
-                        ids.append(cnt)    
-                        cnt += 1
-                        nl += 1
-                fd = os.open(self.cntfn, os.O_EXCL | os.O_TRUNC | os.O_SYNC | os.O_WRONLY)
-                os.write(fd,str(cnt))
-                os.close(fd)
-            except OSError, x:
-                raise RepositoryError("Error on file access: %s" % x)
-            except IOError, x:
-                raise RepositoryError("Error on file access: %s" % x)
-        finally:
-            self.sessionlock.release("counter")
-        self.locked_ids.union_update(Set(ids))
-        return ids
+        return list(self.sessionlock.acquire_next(n))
 
     def lock_ids(self,ids):
         """ lock_ids(n) --> list of successfully locked ids
         Tries to lock the given ids, returning only the successfully locked ids
         Raise RepositoryError"""
-        already_locked_ids = Set()
-        ids_to_lock = Set()
-        for i in ids:
-            assert type(i) == int
-            if i in self.locked_ids:
-                already_locked_ids.add(i)
-            else:
-                ids_to_lock.add(i)
-        for id in ids_to_lock:
-            if self.sessionlock.acquire(str(id)):
-                already_locked_ids.add(id)
-                self.locked_ids.add(id)
-        return list(already_locked_ids)
+        return list(self.sessionlock.acquire(ids))
 
     def release_ids(self,ids):
         """release_ids(ids) --> list of successfully released ids
         Raise RepositoryError"""
-        ids = Set(ids)
-        unlocked_ids = ids.intersection(self.locked_ids)
-        self.locked_ids.difference_update(unlocked_ids)
-        for id in unlocked_ids:
-            self.sessionlock.release(str(id))
-        return list(unlocked_ids)
+        return list(self.sessionlock.release(ids))
 
     def check(self):
-        for id in self.locked_ids:
-            if not os.readlink(os.path.join(self.sessionlock.ldir,str(id))) == self.sessionlock.fn:
-                print "ID ", id, " NOT LOCKED by ", self.sessionlock.fn, " but by ", os.readlink(os.path.join(self.sessionlock.ldir,str(id))), "!!!!!!"
-                assert False
+        self.sessionlock.global_lock()
+        try:
+            f = file(self.sessionlock.cntfn)
+            newcount = int(f.readline())
+            f.close()
+            assert newcount >= self.sessionlock.count
+            sessions = os.listdir(self.sessionlock.sdir)
+            prevnames = Set()
+            for session in sessions:
+                if not session.endswith(".session"):
+                    continue
+                try:
+                    sf = os.path.join(self.sessionlock.sdir,session)
+                    fd = -1
+                    if not self.sessionlock.afs:
+                        fd = os.open(sf, os.O_RDONLY)
+                        fcntl.lockf(fd,fcntl.LOCK_SH) # ONLY NFS
+                    names = pickle.load(file(sf))
+                    if not self.sessionlock.afs and fd > 0:
+                        fcntl.lockf(fd,fcntl.LOCK_UN) # ONLY NFS
+                        os.close(fd)
+                except Exception, x:
+                    logger.warning("CHECKER: session file %s corrupted: %s %s" % (session, x.__class__.__name__, x) )
+                    continue
+                if not len(names & prevnames) == 0:
+                    print "Double-locked stuff:", names & prevnames
+                    assert False
+                prevnames.union_update(names)
+
+        finally:
+            self.sessionlock.global_unlock()
 
 def test1():
     slm = SessionLockManager("locktest","tester")
     while True:
-        print "lock  ---", slm.lock_ids(random.sample(xrange(1000),3))
-        print "unlock---", slm.release_ids(random.sample(xrange(1000),3))
+        print "lock  ---", slm.lock_ids(random.sample(xrange(100),3))
+        print "unlock---", slm.release_ids(random.sample(xrange(100),3))
         slm.check()
 
 def test2():
